Refactor monolithic app into modular structure
